# Conflicts:
#	CHANGELOG.md

+++ tests/unit/test_io_hdf5_streaming.py
@@ -2,7 +2,6 @@ from copy import copy, deepcopy
 import os
 import urllib.request
 import h5py
-import warnings
 
 from hdmf.backends.hdf5.h5tools import HDF5IO
 from hdmf.build import TypeMap, BuildManager
@@ -73,35 +72,29 @@ class TestRos3(TestCase):
         ns_builder.export(self.ns_filename, outdir=self.output_dir)
         ns_path = os.path.join(self.output_dir, self.ns_filename)
 
-        ns_catalog = NamespaceCatalog(NWBGroupSpec, NWBDatasetSpec, NWBNamespace)
+        ns_catalog = NamespaceCatalog(NWBGroupSpec, NWBDatasetSpec, NWBNamespace, core_namespaces=["core"])
         type_map = TypeMap(ns_catalog)
         type_map.merge(get_type_map(), ns_catalog=True)
         type_map.load_namespaces(ns_path)
 
         self.manager = BuildManager(type_map)
 
-        warnings.filterwarnings(action="ignore", message="Ignoring cached namespace .*")
-
     def tearDown(self):
-        if os.path.exists(self.ns_filename):
+        if hasattr(self, 'ns_filename') and os.path.exists(self.ns_filename):
             os.remove(self.ns_filename)
-        if os.path.exists(self.ext_filename):
+        if hasattr(self, 'ext_filename') and os.path.exists(self.ext_filename):
             os.remove(self.ext_filename)
 
+    # NOTE: libhdf5 >= 2.0 made aws_region a required parameter for the ROS3 driver, so all
+    # tests below pass aws_region. See https://github.com/hdmf-dev/hdmf/issues/1470.
+
     def test_basic_read(self):
-        s3_path = "https://dandiarchive.s3.amazonaws.com/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
-
-        with get_hdf5io(s3_path, "r", manager=self.manager, driver="ros3") as io:
-            io.read()
-
-    def test_basic_read_with_aws_region(self):
         s3_path = "https://dandiarchive.s3.amazonaws.com/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
 
         with get_hdf5io(s3_path, "r", manager=self.manager, driver="ros3", aws_region="us-east-2") as io:
             io.read()
 
-    def test_basic_read_s3_with_aws_region(self):
-        # NOTE: if an s3 path is used with ros3 driver, aws_region must be specified
+    def test_basic_read_s3(self):
         s3_path = "s3://dandiarchive/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
 
         with get_hdf5io(s3_path, "r", manager=self.manager, driver="ros3", aws_region="us-east-2") as io:
@@ -111,37 +104,39 @@ class TestRos3(TestCase):
     def test_get_namespaces(self):
         s3_path = "https://dandiarchive.s3.amazonaws.com/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
 
-        namespaces = HDF5IO.get_namespaces(s3_path, driver="ros3")
-        self.assertEqual(namespaces, {'core': '2.3.0', 'hdmf-common': '1.5.0', 'hdmf-experimental': '0.1.0'})
-
-    def test_get_namespaces_with_aws_region(self):
-        s3_path = "https://dandiarchive.s3.amazonaws.com/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
-
         namespaces = HDF5IO.get_namespaces(s3_path, driver="ros3", aws_region="us-east-2")
         self.assertEqual(namespaces, {'core': '2.3.0', 'hdmf-common': '1.5.0', 'hdmf-experimental': '0.1.0'})
 
-    def test_get_namespaces_s3_with_aws_region(self):
+    def test_get_namespaces_s3(self):
         s3_path = "s3://dandiarchive/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
 
         namespaces = HDF5IO.get_namespaces(s3_path, driver="ros3", aws_region="us-east-2")
         self.assertEqual(namespaces, {'core': '2.3.0', 'hdmf-common': '1.5.0', 'hdmf-experimental': '0.1.0'})
+
+    # The setUp loads a stub "core" namespace at version 0.1.0 to mimic PyNWB. The remote test
+    # file caches "core" at 2.3.0, so loading its namespaces emits a UserWarning that the cached
+    # core is being ignored because another version is already loaded.
+    _ignored_core_re = (
+        r"Ignoring the following cached namespace\(s\) because another version is already loaded:"
+        r"\ncore - cached version: 2\.3\.0, loaded version: 0\.1\.0"
+    )
 
     def test_load_namespaces(self):
         s3_path = "https://dandiarchive.s3.amazonaws.com/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
 
-        HDF5IO.load_namespaces(self.manager.namespace_catalog, path=s3_path, driver="ros3")
+        with self.assertWarnsRegex(UserWarning, self._ignored_core_re):
+            HDF5IO.load_namespaces(
+                self.manager.namespace_catalog, path=s3_path, driver="ros3", aws_region="us-east-2"
+            )
         assert set(self.manager.namespace_catalog.namespaces) == set(["core", "hdmf-common", "hdmf-experimental"])
 
-    def test_load_namespaces_with_aws_region(self):
-        s3_path = "https://dandiarchive.s3.amazonaws.com/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
-
-        HDF5IO.load_namespaces(self.manager.namespace_catalog, path=s3_path, driver="ros3", aws_region="us-east-2")
-        assert set(self.manager.namespace_catalog.namespaces) == set(["core", "hdmf-common", "hdmf-experimental"])
-
-    def test_load_namespaces_s3_with_aws_region(self):
+    def test_load_namespaces_s3(self):
         s3_path = "s3://dandiarchive/blobs/11e/c89/11ec8933-1456-4942-922b-94e5878bb991"
 
-        HDF5IO.load_namespaces(self.manager.namespace_catalog, path=s3_path, driver="ros3", aws_region="us-east-2")
+        with self.assertWarnsRegex(UserWarning, self._ignored_core_re):
+            HDF5IO.load_namespaces(
+                self.manager.namespace_catalog, path=s3_path, driver="ros3", aws_region="us-east-2"
+            )
         assert set(self.manager.namespace_catalog.namespaces) == set(["core", "hdmf-common", "hdmf-experimental"])
